+++ sunset/WorldState.hpp
@@ -13,7 +13,7 @@ using std::vector;
 
 struct WorldState {
     u32 currentFrame { 0 };
-    vector<u32> entityIds;
+    vector<u64> entityIds;
     vector<Entity> entities;
     vector<Transform2D> transforms;
     vector<Sprite> sprites;
@@ -21,7 +21,7 @@ struct WorldState {
     vector<Replay> replays;
 
     template<class T>
-    T& GetWorldStateComponent(u32 entityId) {
+    T& GetWorldStateComponent(u64 entityId) {
         if constexpr (std::is_same_v<T, Transform2D>) {
             return transforms.at(FindWorldStateEntityComponent(entityId, ComponentIndex::Transform2D));
         } else if constexpr (std::is_same_v<T, Sprite>) {
@@ -34,12 +34,12 @@ struct WorldState {
     }
 
 private:
-    Entity& FindWorldStateEntity(u32 entityId) {
+    Entity& FindWorldStateEntity(u64 entityId) {
         auto itr = std::lower_bound(entityIds.begin(), entityIds.end(), entityId);
         return entities.at(std::distance(entityIds.begin(), itr));
     }
 
-    i32 FindWorldStateEntityComponent(u32 entityId, ComponentIndex componentIndex) {
+    i32 FindWorldStateEntityComponent(u64 entityId, ComponentIndex componentIndex) {
         return FindWorldStateEntity(entityId).components.at(static_cast<i32>(componentIndex));
     }
 };
